problems

+++ windows/practice/G.cpp
@@ -1,13 +1,11 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-
 int main() {
 	ios::sync_with_stdio(0); cin.tie(0);
-	int tc = 1; 
+	long long tc = 1; 
 	// cin >> tc;
-	for (int tt = 0; tt < tc; tt++) {
+	for (long long tt = 0; tt < tc; tt++) {
 		
- 
 	}
 }
